[#2] Refactor: components 하위 폴더 구조 변경
