+++ src/pages/Root/index.tsx
@@ -1,7 +1,7 @@
 import { Box } from '@chakra-ui/react';
 import { Outlet } from 'react-router-dom';
 
-import { Navbar } from '@/components/Navbar';
+import { Navbar } from '@/components/common/Navbar';
 
 export const Root = () => {
   return (
